refactor: split up app view
