+++ crates/private-server/src/app/greeting.rs
@@ -1,0 +1,12 @@
+use leptos::prelude::*;
+
+#[component]
+pub fn Greeting() -> impl IntoView {
+	let greeting = crate::statuses::greeting();
+
+	view! {
+		<Await future=greeting let:data>
+			<div class="greeting">{data.clone().ok()}</div>
+		</Await>
+	}
+}
